# src/main/resources/api/neoris-api-swagger.yaml
openapi: 3.0.3
info:
  title: Neoris - Price API
  description: |-
    This is the description of the API, describe the way to access and different methods exposed in the API.
  contact:
    email: danieltrujilloceron@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8080/neoris/api/price
tags:
  - name: price
    description: all service about your prices
paths:
  /product/{productId}/brand/{brandId}:
    get:
      tags:
        - price
      summary: Find Price by productId and brandId
      description: Returns a single price
      operationId: getPriceByProductIdAndBrandIdAndDateBetweenStartDateAndEndDateAndFilterByMajorPriority
      parameters:
        - name: productId
          in: path
          description: ID of product
          required: true
          schema:
            type: integer
            format: int64
        - name: brandId
          in: path
          description: ID of brand
          required: true
          schema:
            type: integer
            format: int64
        - name: application-date
          in: query
          description: Status values that need to be considered for filter
          required: true
          explode: true
          schema:
            type: string
            default: "2020-06-16T21:00:00"
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/price'
components:
  schemas:
    price:
      type: object
      properties:
        priceList:
          type: integer
          format: int64
          example: 10
        brandId:
          type: integer
          format: int64
          example: 10
        startDate:
          type: string
          format: date-time
          example: "2020-06-16T21:00:00"
        endDate:
          type: string
          format: date-time
          example: "2020-06-16T21:00:00"
        productId:
          type: integer
          format: int64
          example: 198772
        priority:
          type: integer
          format: int64
          example: 1
        price:
          type: number
          format: double
          example: 35.5
        currency:
          type: string
          format: date-time
          example: "EUR"